backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyserial
scikit-learn
pandas
numpy
python-dotenv
openai
sqlalchemy
alembic
websockets
